Restore ElectionContext.elgamal_public_key name to match python. Cleanup Json Pojos. Move serialize and deserialize into ElectionDescriptionPojo.
