--- src/main/java/com/sunya/electionguard/publish/ElectionDescriptionFromJson.java
@@ -1,185 +1,0 @@
-package com.sunya.electionguard.publish;
-
-import com.google.common.base.Strings;
-import com.google.common.flogger.FluentLogger;
-import com.google.gson.Gson;
-import com.google.gson.JsonElement;
-import com.sunya.electionguard.Election;
-
-import javax.annotation.Nullable;
-import java.io.*;
-import java.time.OffsetDateTime;
-import java.util.List;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-
-/** Helper class for conversion of ElectionDescription from Json. */
-public class ElectionDescriptionFromJson {
-  private static final FluentLogger logger = FluentLogger.forEnclosingClass();
-
-  public static Election deserialize(JsonElement jsonElem) {
-    Gson gson = GsonTypeAdapters.enhancedGson();
-    ElectionDescriptionPojo pojo = gson.fromJson(jsonElem, ElectionDescriptionPojo.class);
-    ElectionDescriptionFromJson converter = new ElectionDescriptionFromJson(null);
-    return converter.convert(pojo);
-  }
-
-  private final String pathname;
-  public ElectionDescriptionFromJson(String pathname) {
-    this.pathname = pathname;
-  }
-
-  // used by consumer
-  public Election build() throws IOException {
-    try (InputStream is = new FileInputStream(pathname)) {
-      Reader reader = new InputStreamReader(is);
-      Gson gson = new Gson(); // default exclude nulls
-      ElectionDescriptionPojo pojo = gson.fromJson(reader, ElectionDescriptionPojo.class);
-      return convert(pojo);
-    } catch (Exception ioe) {
-      logger.atSevere().log("Failed on file '%s'", pathname);
-      throw ioe;
-    }
-  }
-
-  @Nullable
-  <T, U> List<U> convertList(@Nullable List<T> from, Function<T, U> converter) {
-    return from == null || from.isEmpty() ? null : from.stream().map(converter).collect(Collectors.toList());
-  }
-
-  Election convert(ElectionDescriptionPojo pojo) {
-    // LOOK going to throw an Exception if these arent present or invalid format
-    OffsetDateTime startLocalDate = OffsetDateTime.parse(pojo.start_date);
-    OffsetDateTime endLocalDate = OffsetDateTime.parse(pojo.end_date);
-
-    return new Election(
-            pojo.election_scope_id,
-            Election.ElectionType.valueOf(pojo.type),
-            startLocalDate,
-            endLocalDate,
-            convertList(pojo.geopolitical_units, this::convertGeopoliticalUnit),
-            convertList(pojo.parties, this::convertParty),
-            convertList(pojo.candidates, this::convertCandidate),
-            convertList(pojo.contests, this::convertContestDescription),
-            convertList(pojo.ballot_styles, this::convertBallotStyle),
-            convertInternationalizedText(pojo.name),
-            convertContactInformation(pojo.contact_information)
-    );
-  }
-
-  @Nullable
-  Election.AnnotatedString convertAnnotatedString(@Nullable ElectionDescriptionPojo.AnnotatedString pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.AnnotatedString(
-            Strings.nullToEmpty(pojo.annotation),
-            Strings.nullToEmpty(pojo.value));
-  }
-
-
-  @Nullable
-  Election.BallotStyle convertBallotStyle(@Nullable ElectionDescriptionPojo.BallotStyle pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.BallotStyle(pojo.object_id,
-            convertList(pojo.geopolitical_unit_ids, Strings::nullToEmpty),
-            convertList(pojo.party_ids, Strings::nullToEmpty),
-            Strings.emptyToNull(pojo.image_uri));
-  }
-
-  @Nullable
-  Election.Candidate convertCandidate(@Nullable ElectionDescriptionPojo.Candidate pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.Candidate(pojo.object_id, convertInternationalizedText(pojo.name),
-            pojo.party_id, pojo.image_uri, pojo.is_write_in);
-  }
-
-  @Nullable
-  Election.ContactInformation convertContactInformation(@Nullable ElectionDescriptionPojo.ContactInformation pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.ContactInformation(
-            convertList(pojo.address_line, Strings::nullToEmpty),
-            convertList(pojo.email, this::convertAnnotatedString),
-            convertList(pojo.phone, this::convertAnnotatedString),
-            pojo.name);
-  }
-
-  @Nullable
-  Election.ContestDescription convertContestDescription(@Nullable ElectionDescriptionPojo.ContestDescription pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.ContestDescription(
-            pojo.object_id,
-            pojo.electoral_district_id,
-            pojo.sequence_order,
-            Election.VoteVariationType.valueOf(pojo.vote_variation),
-            pojo.number_elected,
-            pojo.votes_allowed,
-            pojo.name,
-            convertList(pojo.ballot_selections, this::convertSelectionDescription),
-            convertInternationalizedText(pojo.ballot_title),
-            convertInternationalizedText(pojo.ballot_subtitle));
-  }
-
-  @Nullable
-  Election.GeopoliticalUnit convertGeopoliticalUnit(@Nullable ElectionDescriptionPojo.GeopoliticalUnit pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.GeopoliticalUnit(
-            pojo.object_id,
-            pojo.name,
-            Election.ReportingUnitType.valueOf(pojo.type),
-            convertContactInformation(pojo.contact_information));
-  }
-
-  @Nullable
-  Election.InternationalizedText convertInternationalizedText(@Nullable ElectionDescriptionPojo.InternationalizedText pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.InternationalizedText(convertList(pojo.text, this::convertLanguage));
-  }
-
-  @Nullable
-  Election.Language convertLanguage(@Nullable ElectionDescriptionPojo.Language pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.Language(
-            Strings.nullToEmpty(pojo.value),
-            Strings.nullToEmpty(pojo.language));
-  }
-
-  @Nullable
-  Election.Party convertParty(@Nullable ElectionDescriptionPojo.Party pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.Party(
-            pojo.object_id,
-            convertInternationalizedText(pojo.ballot_name),
-            pojo.abbreviation,
-            pojo.color,
-            pojo.logo_uri);
-  }
-
-  @Nullable
-  Election.SelectionDescription convertSelectionDescription(@Nullable ElectionDescriptionPojo.SelectionDescription pojo) {
-    if (pojo == null) {
-      return null;
-    }
-    return new Election.SelectionDescription(
-            pojo.object_id,
-            pojo.candidate_id,
-            pojo.sequence_order);
-  }
-
-}
